html/semantics/forms/the-button-element/button-type-enumerated-ascii-case-insensitive.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="help" href="https://html.spec.whatwg.org/#attr-button-type">
<link rel="help" href="https://html.spec.whatwg.org/#enumerated-attribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<button type="reset">
<button type="ReSeT">
<button type="reſet">
<button type="submit">
<button type="SuBmIt">
<button type="ſubmit">
<script>
test(() => {
  const button = document.querySelectorAll("button");

  assert_equals(button[0].type, "reset");
  assert_equals(button[1].type, "reset");
  assert_equals(button[2].type, "submit");
  assert_equals(button[3].type, "submit");

  // vacuous: the invalid value default is currently submit, so even if the UA
  // treats this as invalid, the observable behaviour would still be correct
  assert_equals(button[4].type, "submit");

  // vacuous: the invalid value default is currently submit, so even if the UA
  // treats this as valid, the observable behaviour would still be correct
  assert_equals(button[5].type, "submit");
});
</script>
